example-2.1/requirements.txt
requests
azure-storage-blob
python-dateutil
pytz
timedinput